refactor: move raster definitions to own module
