+++ tests/unittests/test_raster.py
@@ -6,7 +6,7 @@ from unittest import mock
 import geojson
 
 import ohsome_quality_analyst.raster.client as raster_client
-from ohsome_quality_analyst.definitions import get_raster_dataset
+from ohsome_quality_analyst.raster.definitions import get_raster_dataset
 from ohsome_quality_analyst.utils.exceptions import RasterDatasetNotFoundError
