+++ tests/unittests/test_raster_definitions.py
@@ -1,0 +1,19 @@
+import unittest
+
+from ohsome_quality_analyst.raster import definitions
+from ohsome_quality_analyst.utils.exceptions import RasterDatasetUndefinedError
+
+
+class TestDefinitions(unittest.TestCase):
+    def test_get_raster_dataset_names(self):
+        names = definitions.get_raster_dataset_names()
+        self.assertIsInstance(names, list)
+        self.assertTrue(names)
+
+    def test_get_raster_dataset(self):
+        raster = definitions.get_raster_dataset("GHS_BUILT_R2018A")
+        self.assertIsInstance(raster, definitions.RasterDataset)
+
+    def test_get_raster_dataset_undefined(self):
+        with self.assertRaises(RasterDatasetUndefinedError):
+            definitions.get_raster_dataset("foo")
